+++ parser_library/src/expressions/conditional_assembly/terms/ca_symbol_attribute.cpp
@@ -15,12 +15,18 @@
 #include "ca_symbol_attribute.h"
 
 #include "ca_var_sym.h"
+#include "context/literal_pool.h"
 #include "context/ordinary_assembly/dependable.h"
 #include "context/ordinary_assembly/ordinary_assembly_dependency_solver.h"
 #include "ebcdic_encoding.h"
 #include "expressions/conditional_assembly/ca_expr_visitor.h"
 #include "expressions/evaluation_context.h"
+#include "hlasmparser.h"
 #include "lexing/lexer.h"
+#include "lexing/token_stream.h"
+#include "parsing/parser_impl.h"
+#include "processing/op_code.h"
+#include "semantics/range_provider.h"
 #include "semantics/statement_fields.h"
 
 namespace hlasm_plugin::parser_library::expressions {
@@ -240,27 +246,21 @@ context::SET_t ca_symbol_attribute::evaluate_ordsym(context::id_index name, cons
 context::SET_t ca_symbol_attribute::evaluate_literal(
     const semantics::literal_si& lit, const evaluation_context& eval_ctx) const
 {
-    context::ordinary_assembly_dependency_solver solver(
-        eval_ctx.hlasm_ctx.ord_ctx, eval_ctx.hlasm_ctx.ord_ctx.align(context::no_align));
+    auto& literals = eval_ctx.hlasm_ctx.ord_ctx.literals();
 
     if (attribute == context::data_attr_kind::D)
-        return false;
-    else if (attribute == context::data_attr_kind::O)
-    {
-        eval_ctx.diags.add_diagnostic(diagnostic_op::error_E066(expr_range));
-        return {};
-    }
-    else if (attribute == context::data_attr_kind::T)
-    {
+        return literals.defined_for_ca_expr(std::shared_ptr<const expressions::data_definition>(lit, &lit->get_dd()));
+
+    literals.mentioned_in_ca_expr(std::shared_ptr<const expressions::data_definition>(lit, &lit->get_dd()));
+
+    if (attribute == context::data_attr_kind::O)
         return "U";
-    }
+    else if (attribute == context::data_attr_kind::T)
+        return std::string { lit->get_dd().get_type_attribute() };
     else
     {
-        context::symbol_attributes attrs(context::symbol_origin::DAT,
-            ebcdic_encoding::a2e[(unsigned char)lit->get_dd().get_type_attribute()],
-            lit->get_dd().get_length_attribute(solver, eval_ctx.diags),
-            lit->get_dd().get_scale_attribute(solver, eval_ctx.diags),
-            lit->get_dd().get_integer_attribute(solver, eval_ctx.diags));
+        context::ordinary_assembly_dependency_solver solver(eval_ctx.hlasm_ctx.ord_ctx, context::address());
+        context::symbol_attributes attrs = lit->get_dd().get_symbol_attributes(solver, eval_ctx.diags);
         if ((attribute == context::data_attr_kind::S || attribute == context::data_attr_kind::I)
             && !attrs.can_have_SI_attr())
         {
@@ -335,8 +335,12 @@ context::SET_t ca_symbol_attribute::evaluate_substituted(context::id_index var_n
         return context::symbol_attributes::default_ca_value(attribute);
     }
 
-    auto [valid, ord_name] =
-        eval_ctx.hlasm_ctx.try_get_symbol_name(try_extract_leading_symbol(substituted_name.access_c()));
+    const auto& text = substituted_name.access_c();
+
+    if (!text.empty() && text.starts_with('='))
+        return evaluate_substituted_literal(text, var_range, eval_ctx);
+
+    auto [valid, ord_name] = eval_ctx.hlasm_ctx.try_get_symbol_name(try_extract_leading_symbol(text));
 
     if (!valid)
     {
@@ -346,6 +350,50 @@ context::SET_t ca_symbol_attribute::evaluate_substituted(context::id_index var_n
     }
     else
         return evaluate_ordsym(ord_name, eval_ctx);
+}
+
+context::SET_t ca_symbol_attribute::evaluate_substituted_literal(
+    const std::string& text, range var_range, const evaluation_context& eval_ctx) const
+{
+    // error production is suppressed when evaluating D', T' and O' attributes
+    using attr_kind = context::data_attr_kind;
+    const bool suppress = attribute == attr_kind::T || attribute == attr_kind::D || attribute == attr_kind::O;
+    bool error = false;
+    diagnostic_consumer_transform add_diag_subst([&text, &eval_ctx, suppress, &error](diagnostic_op diag) {
+        error = true;
+        if (suppress)
+            return;
+        diag.message = diagnostic_decorate_message(text, diag.message);
+        eval_ctx.diags.add_diagnostic(std::move(diag));
+    });
+    auto h = parsing::parser_holder::create(nullptr, &eval_ctx.hlasm_ctx, &add_diag_subst);
+
+    h->input->reset(text);
+
+    h->lex->reset();
+    h->lex->set_file_offset(var_range.start);
+    h->lex->set_unlimited_line(true);
+
+    h->stream->reset();
+
+    h->parser->reinitialize(&eval_ctx.hlasm_ctx,
+        semantics::range_provider(var_range, semantics::adjusting_state::SUBSTITUTION),
+        processing::processing_status(processing::processing_format(processing::processing_kind::ORDINARY,
+                                          processing::processing_form::CA,
+                                          processing::operand_occurence::ABSENT),
+            processing::op_code()),
+        &add_diag_subst);
+
+    h->parser->reset();
+
+    h->parser->get_collector().prepare_for_next_statement();
+
+    auto literal_context = h->parser->literal_reparse();
+
+    if (!error && literal_context->value)
+        return evaluate_literal(literal_context->value, eval_ctx);
+    else
+        return context::symbol_attributes::default_ca_value(attribute);
 }
 
 } // namespace hlasm_plugin::parser_library::expressions