+++ parser_library/src/expressions/conditional_assembly/terms/ca_symbol_attribute.h
@@ -67,6 +67,8 @@ private:
         range var_range,
         const evaluation_context& eval_ctx) const;
     context::SET_t evaluate_literal(const semantics::literal_si& lit, const evaluation_context& eval_ctx) const;
+    context::SET_t evaluate_substituted_literal(
+        const std::string& text, range var_range, const evaluation_context& eval_ctx) const;
 };
 
 } // namespace hlasm_plugin::parser_library::expressions